Adding args: port & no-pid

+++ cml.py
@@ -1,5 +1,5 @@
 #!/bin/python3
-import os,  socketserver, hashlib, urllib.request
+import os,  socketserver
 import config
 import utils
 import handler
@@ -13,11 +13,13 @@ class Cml:
         u = utils.Main(s.ARGS.verbosity)
 
         # Pid check
-        pid = str(os.getpid())
-        if os.path.isfile(s.CONFIG["SERVER"]["pidfile"]):
-            u.print(s.CONFIG["SERVER"]["pidfile"],"already exists, exiting")
-            exit()
-        else: open(s.CONFIG["SERVER"]["pidfile"], 'w').write(pid)
+        if not s.ARGS.no_pid:
+            pid = str(os.getpid())
+            if os.path.isfile(s.CONFIG["SERVER"]["pidfile"]):
+                u.print(s.CONFIG["SERVER"]["pidfile"],
+                        "already exists, exiting")
+                exit()
+            else: open(s.CONFIG["SERVER"]["pidfile"], 'w').write(pid)
 
         # Cache check
         if not os.path.exists(s.CONFIG["CACHE"]["cache-dir"]):
@@ -26,12 +28,15 @@ class Cml:
             os.mkdir(s.CONFIG["CACHE"]["cache-dir"])
 
         # MAIN
-        PORT = int(s.CONFIG["SERVER"]["port"])
+        if s.ARGS.port: PORT = int(s.ARGS.port)
+        else: PORT = int(s.CONFIG["SERVER"]["port"])
+
         try:
             httpd = socketserver.TCPServer(("", PORT), handler.MyHandler)
         except OSError:
             u.print("Error: ip already binded")
-            os.unlink(s.CONFIG["SERVER"]["pidfile"])
+            if not s.ARGS.no_pid:
+                os.unlink(s.CONFIG["SERVER"]["pidfile"])
             exit()
 
         u.print("serving at port", PORT)
@@ -39,7 +44,8 @@ class Cml:
             httpd.serve_forever()
         except KeyboardInterrupt:
             u.print("exiting...")
-            os.unlink(s.CONFIG["SERVER"]["pidfile"])
+            if not s.ARGS.no_pid:
+                os.unlink(s.CONFIG["SERVER"]["pidfile"])
 cml = Cml()
